Fix propagation of errors in `multiprocessing.util.spawnv_passfds`
The error return pipe, `errpipe_write`, was not included in the "FDs to
keep" that were passed to `_posixsubprocess.fork_exec`. This resulted in
the pipe being closed in `_posixsubprocess.child_exec`, which in turn
meant errors from `os.execv`/`os.execve` could not be written to it,
preventing propagation.

We saw this manifest when our env vars became too large. `strace`
showed:
```
[pid   114] execve("/default-pegasus-venv/bin/python", ["/default-pegasus-venv/bin/python", "-B", "-c", "from multiprocessing.resource_tr"...], 0x62a996e37620 /* 311 vars */) = -1 E2BIG (Argument list too long)
[pid   114] write(31, "OSError:", 8)    = -1 EBADF (Bad file descriptor)
[pid   114] write(31, "7", 1)           = -1 EBADF (Bad file descriptor)
[pid   114] write(31, ":", 1)           = -1 EBADF (Bad file descriptor)
[pid   114] write(31, "", 0)            = -1 EBADF (Bad file descriptor)
```

This change fixes this, matching the behaviour in
`subprocess.Popen._execute_child`:
https://github.com/python/cpython/blob/138ed6db9f89171983dc32af4e7ad2e73d46a940/Lib/subprocess.py#L1897-L1901